+++ CHANGELOG.md
@@ -18,6 +18,7 @@ it's set lower
 * Fixed game ending message was never showed when using any of the locales (accessor was broken)
 * /bba reload command was undeprecated and it's usage is no longer discouraged
 * Made /bb help and /bba help commands actually working
+* Added not playing message node to language.yml, locales already had it somehow
 
 ### 4.0.0 Beta Pre Releases 3-15 (03.10.2018 - 06.06.2019)
    > RC 2